스케줄러, 대관하기 페이지 연동
